Extracted CollectionTask.UpdateValuesFromDeck to updateValuesFromDeck()
RebuildCram and EmptyCram calls directly UpdateValuesFromDeck.executeTask() which in-turn calls UpdateValuesFromDeck.task() directly in a blocking way. Directly calling updateValuesFromDeck() is equivalent.
